Project_Unthinkable/requirements.txt
google-cloud-speech==2.33.0
pydub
ffmpeg-python
whisper
transformers
torch
torchaudio
pyannote.audio
